+++ src/mprTest.c
@@ -877,7 +877,7 @@ bool mprWaitForTestToComplete(MprTestGroup *gp, int timeout)
     expires = mprGetTime() + timeout;
     remaining = timeout;
     do {
-        mprWaitForEvent(gp->dispatcher, (int) remaining);
+        mprWaitForEvent(gp->dispatcher, remaining);
         remaining = expires - mprGetTime();
     } while (!gp->testComplete && remaining > 0);
     rc = gp->testComplete;
